# api/openapi.yaml
openapi: 3.1.0
info:
  title: ILTER Gateway API
  description: Secure, cost-optimizing LLM Gateway with built-in PII masking, loop detection, and load balancing.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:9191
    description: Dashboard Server
paths:
...
  /v1/chat/completions:
    post:
      tags: [Proxy]
      servers:
        - url: http://127.0.0.1:8181
      summary: Proxy Chat Completion
      description: Proxies chat completions to the load-balanced providers.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                model:
                  type: string
                  example: gpt-4o
                messages:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant, tool]
                        example: user
                      content:
                        type: string
                        example: Hello, how are you?
                stream:
                  type: boolean
                  example: false
                temperature:
                  type: number
                  minimum: 0.0
                  maximum: 2.0
                  example: 0.7
                max_tokens:
                  type: integer
                  minimum: 1
                  example: 1024
      responses:
        '200':
          description: Completion successful
          headers:
            X-Ilter-Complexity-Score:
              schema:
                type: number
                description: Complexity score of the prompt (0.0–10.0)
            X-Ilter-Model-Selected:
              schema:
                type: string
                description: The model selected by smart routing
            X-Ilter-Cost:
              schema:
                type: number
                description: Actual cost in USD for this request (non-streaming only)
            X-Ilter-Cost-Estimate:
              schema:
                type: number
                description: Estimated cost in USD for this request
            X-Ilter-Alternative-Cost:
              schema:
                type: number
                description: Estimated cost if using an alternative model
            X-Ilter-Savings-Potential:
              schema:
                type: string
                description: Savings percentage compared to alternative
                example: 45%
            X-Ilter-Loop-Warning:
              schema:
                type: boolean
                description: Indicates potential agentic loop detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  object:
                    type: string
                  created:
                    type: integer
                  model:
                    type: string
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        message:
                          type: object
                          properties:
                            role:
                              type: string
                            content:
                              type: string
                        finish_reason:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/PIIBlocked'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/ProviderError'
...
  /webhooks/{token}:
    post:
      tags: [Admin]
      servers:
        - url: http://127.0.0.1:9191
      summary: Webhook Entry
      description: |
        External unauthenticated endpoint for receiving webhook events.
        HMAC verification is performed using SHA-256.
        Returns 202 Accepted with run_id on success, or 401 on HMAC failure.
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              description: Raw webhook payload
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            description: Webhook token (wht_ prefix)
      responses:
        '202':
          description: Webhook accepted — polling endpoint (use run_id with GET /api/jobs/{id}/runs/{runId})
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    description: The run ID to poll for completion
        '401':
          description: HMAC verification failed or trigger not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Trigger not found or verification failed
                      type:
                        type: string
                        example: verification_failed
  /api/webhooks/{token}:
    post:
      tags: [Dashboard]
      servers:
        - url: http://127.0.0.1:9191
      summary: Webhook Entry
      description: |
        External unauthenticated endpoint for receiving webhook events.
        HMAC verification is performed using SHA-256.
        Returns 202 Accepted with run_id on success, or 401 on HMAC failure.
      requestBody:
        required: true
        content:
          '*/*':
            schema:
              type: string
              description: Raw webhook payload
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            description: Webhook token (wht_ prefix)
      responses:
        '202':
          description: Webhook accepted — polling endpoint (use run_id with GET /api/jobs/{id}/runs/{runId})
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    description: The run ID to poll for completion
        '401':
          description: HMAC verification failed or trigger not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Trigger not found or verification failed
                      type:
                        type: string
                        example: verification_failed


  # ---------------------------------------------------------------------------
  # OAuth 2.0 — MCP Auth endpoints (RFC 8414, RFC 8707, RFC 7591, PKCE)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    BadRequest:
      description: Malformed request payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    PIIBlocked:
      description: Request blocked due to PII detection
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  pii_type:
                    type: string
                    description: Type of PII detected (e.g., 'ssn', 'email', 'credit_card', 'tckn')
                  mode:
                    type: string
                    description: The PII mode that triggered the block
                    example: block
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    example: model_not_found
                  code:
                    type: string
                    example: "404"
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    example: rate_limit_exceeded
                  code:
                    type: string
                    example: "429"
    ProviderError:
      description: All upstream providers failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    example: provider_error
                  code:
                    type: string
                    example: "502"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                    example: internal_error
                  code:
                    type: string
                    example: "500"
...